+++ RandomUsers/RandomUsers/RandomUsers.csproj
@@ -48,11 +48,14 @@
     <Compile Include="Models\Result.cs" />
     <Compile Include="Models\User.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="UserDataTemplateSelector.cs" />
     <Compile Include="UserDetail.xaml.cs">
       <DependentUpon>UserDetail.xaml</DependentUpon>
     </Compile>
     <Compile Include="Service\DataService.cs" />
     <Compile Include="ViewModel\MainPageViewModel.cs" />
+    <Compile Include="ViewModel\PageViewModelBase.cs" />
+    <Compile Include="ViewModel\UserDetailViewModel.cs" />
     <Compile Include="ViewModel\ViewModelLocator.cs" />
   </ItemGroup>
   <ItemGroup>
