+++ RandomUsers/RandomUsers/ViewModel/PageViewModelBase.cs
@@ -1,0 +1,17 @@
+﻿using GalaSoft.MvvmLight;
+using Xamarin.Forms;
+
+namespace RandomUsers.ViewModel
+{
+
+    public class PageViewModelBase : ViewModelBase
+    {
+        protected INavigation Navigation
+        {
+            get
+            {
+                return Application.Current.MainPage.Navigation;
+            }
+        }
+    }
+}
