+++ RandomUsers/RandomUsers/ViewModel/ViewModelLocator.cs
@@ -1,5 +1,7 @@
 ﻿using GalaSoft.MvvmLight.Ioc;
+using GalaSoft.MvvmLight.Messaging;
 using Microsoft.Practices.ServiceLocation;
+using RandomUsers.Models;
 using RandomUsers.ViewModel;
 using System;
 using System.Collections.Generic;
@@ -15,6 +17,7 @@ namespace RandomUsers
         {
             ServiceLocator.SetLocatorProvider(() => SimpleIoc.Default);
             SimpleIoc.Default.Register<MainPageViewModel>();
+            SimpleIoc.Default.Register<UserDetailViewModel>();
         }
 
         public MainPageViewModel Main
@@ -22,6 +25,14 @@ namespace RandomUsers
             get
             {
                 return ServiceLocator.Current.GetInstance<MainPageViewModel>();
+            }
+        }
+
+        public UserDetailViewModel Detail
+        {
+            get
+            {
+                return ServiceLocator.Current.GetInstance<UserDetailViewModel>();
             }
         }
     }
